+++ delphi-frontend/src/test/java/au/com/integradev/delphi/antlr/GrammarTest.java
@@ -371,4 +371,9 @@ class GrammarTest {
   void testGenericConstraints() {
     assertParsed("GenericConstraints.pas");
   }
+
+  @Test
+  void testIfExpression() {
+    assertParsed("IfExpression.pas");
+  }
 }