+++ delphi-frontend/src/test/java/au/com/integradev/delphi/symbol/resolve/ExpressionTypeResolverTest.java
@@ -1,0 +1,91 @@
+/*
+ * Sonar Delphi Plugin
+ * Copyright (C) 2019 Integrated Application Development
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
+ */
+package au.com.integradev.delphi.symbol.resolve;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+import au.com.integradev.delphi.DelphiProperties;
+import au.com.integradev.delphi.compiler.Toolchain;
+import au.com.integradev.delphi.type.factory.TypeFactoryImpl;
+import org.junit.jupiter.api.Test;
+import org.sonar.plugins.communitydelphi.api.ast.ExpressionNode;
+import org.sonar.plugins.communitydelphi.api.ast.IfExpressionNode;
+import org.sonar.plugins.communitydelphi.api.type.IntrinsicType;
+import org.sonar.plugins.communitydelphi.api.type.Type;
+import org.sonar.plugins.communitydelphi.api.type.TypeFactory;
+
+class ExpressionTypeResolverTest {
+  private static final TypeFactory TYPE_FACTORY =
+      new TypeFactoryImpl(Toolchain.DCC64, DelphiProperties.COMPILER_VERSION_DEFAULT);
+  private static final ExpressionTypeResolver RESOLVER = new ExpressionTypeResolver(TYPE_FACTORY);
+
+  @Test
+  void testIdenticalTypesResolveToThatType() {
+    Type integer = TYPE_FACTORY.getIntrinsic(IntrinsicType.INTEGER);
+    assertThat(resolveIfExpression(integer, integer)).isEqualTo(integer);
+  }
+
+  @Test
+  void testIntegerAndRealResolveToReal() {
+    Type integer = TYPE_FACTORY.getIntrinsic(IntrinsicType.INTEGER);
+    Type extended = TYPE_FACTORY.getIntrinsic(IntrinsicType.EXTENDED);
+    assertThat(resolveIfExpression(integer, extended)).isEqualTo(extended);
+    assertThat(resolveIfExpression(extended, integer)).isEqualTo(extended);
+  }
+
+  @Test
+  void testNarrowerIntegerWidensToWiderIntegerRegardlessOfBranchOrder() {
+    Type shortInt = TYPE_FACTORY.getIntrinsic(IntrinsicType.SHORTINT);
+    Type int64 = TYPE_FACTORY.getIntrinsic(IntrinsicType.INT64);
+    assertThat(resolveIfExpression(shortInt, int64)).isEqualTo(int64);
+    assertThat(resolveIfExpression(int64, shortInt)).isEqualTo(int64);
+  }
+
+  @Test
+  void testIncompatibleTypesResolveToUnknown() {
+    Type integer = TYPE_FACTORY.getIntrinsic(IntrinsicType.INTEGER);
+    Type string = TYPE_FACTORY.getIntrinsic(IntrinsicType.UNICODESTRING);
+    assertThat(resolveIfExpression(integer, string).isUnknown()).isTrue();
+  }
+
+  @Test
+  void testUnknownBranchTypeResolvesToUnknown() {
+    Type integer = TYPE_FACTORY.getIntrinsic(IntrinsicType.INTEGER);
+    assertThat(resolveIfExpression(integer, TypeFactory.unknownType()).isUnknown()).isTrue();
+    assertThat(resolveIfExpression(TypeFactory.unknownType(), integer).isUnknown()).isTrue();
+  }
+
+  private static Type resolveIfExpression(Type thenType, Type elseType) {
+    ExpressionNode thenExpression = mockExpressionNode(thenType);
+    ExpressionNode elseExpression = mockExpressionNode(elseType);
+
+    IfExpressionNode ifExpression = mock(IfExpressionNode.class);
+    when(ifExpression.getThenExpression()).thenReturn(thenExpression);
+    when(ifExpression.getElseExpression()).thenReturn(elseExpression);
+    return RESOLVER.resolve(ifExpression);
+  }
+
+  private static ExpressionNode mockExpressionNode(Type type) {
+    ExpressionNode result = mock();
+    when(result.getType()).thenReturn(type);
+    return result;
+  }
+}