+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for `if` conditional expressions, introduced in Delphi 13.
+
 ## [1.19.0] - 2026-06-09
 
 ### Added